Move exception classes to a sub-namespace.
