+++ src/Exception/ResourceNotFoundException.php
@@ -2,7 +2,7 @@
 
 /* vim: set expandtab tabstop=4 shiftwidth=4: */
 
-namespace silverorange\JsonApiClient;
+namespace silverorange\JsonApiClient\Exception;
 
 class ResourceNotFoundExeption extends ResourceErrorException
 {
